Stock/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.43.0",
    "websocket>=0.2.1",
    "websocket-client>=1.8.0",
    "yfinance>=0.2.54",
]
